pyproject.toml
[tool.poetry]
name = "owl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
protobuf = "^5.26.1"
cpm-kernels = "^1.0.11"
torch = "^2.2.2"
gradio = "^4.24.0"
mdtex2html = "^1.3.0"
sentencepiece = "^0.2.0"
accelerate = "^0.28.0"
datasets = "^2.18.0"
matplotlib = "^3.8.3"
huggingface-hub = "^0.22.2"
triton = "^2.2.0"
transformers = "^4.39.2"
tiktoken = "^0.6.0"
einops = "^0.7.0"
transformers-stream-generator = "^0.0.5"
scipy = "^1.12.0"
maritalk = "^0.2.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.2"
groq = "^0.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
